+++ source/Energinet.DataHub.MarketParticipant.IntegrationTests/Repositories/GridAreaAuditLogEntryRepositoryTests.cs
@@ -16,6 +16,7 @@ using System;
 using System.Linq;
 using System.Threading.Tasks;
 using Energinet.DataHub.MarketParticipant.Domain.Model;
+using Energinet.DataHub.MarketParticipant.Domain.Model.Users;
 using Energinet.DataHub.MarketParticipant.Infrastructure.Persistence.Repositories;
 using Energinet.DataHub.MarketParticipant.IntegrationTests.Fixtures;
 using Microsoft.EntityFrameworkCore;
@@ -46,25 +47,25 @@ namespace Energinet.DataHub.MarketParticipant.IntegrationTests.Repositories
             var gridAreaRepo = new GridAreaRepository(context);
             var gridAreaId = await gridAreaRepo.AddOrUpdateAsync(new GridArea(new GridAreaName("name"), new GridAreaCode("1234"), PriceAreaCode.Dk1));
 
-            var userId = Guid.NewGuid();
+            var externalUserId = new ExternalUserId(Guid.NewGuid());
 
             var target = new GridAreaAuditLogEntryRepository(context);
-            var entry = new GridAreaAuditLogEntry(DateTimeOffset.UtcNow, userId, GridAreaAuditLogEntryField.Name, "old_val", "new_val", gridAreaId.Value);
+            var entry = new GridAreaAuditLogEntry(DateTimeOffset.UtcNow, externalUserId, GridAreaAuditLogEntryField.Name, "old_val", "new_val", gridAreaId);
 
             // act
             await target.InsertAsync(entry);
 
             var actual = await (from l in context.GridAreaAuditLogEntries.AsQueryable()
-                                where l.UserId == userId
+                                where l.UserId == externalUserId.Value
                                 select l).SingleAsync();
 
             // assert
             Assert.Equal(entry.Timestamp, actual.Timestamp);
-            Assert.Equal(entry.UserId, actual.UserId);
+            Assert.Equal(entry.ExternalUserId.Value, actual.UserId);
             Assert.Equal(entry.Field, actual.Field);
             Assert.Equal(entry.OldValue, actual.OldValue);
             Assert.Equal(entry.NewValue, actual.NewValue);
-            Assert.Equal(entry.GridAreaId, actual.GridAreaId);
+            Assert.Equal(entry.GridAreaId.Value, actual.GridAreaId);
         }
 
         [Fact]
@@ -78,10 +79,10 @@ namespace Energinet.DataHub.MarketParticipant.IntegrationTests.Repositories
             var gridAreaRepo = new GridAreaRepository(context);
             var gridAreaId = await gridAreaRepo.AddOrUpdateAsync(new GridArea(new GridAreaName("name"), new GridAreaCode("1234"), PriceAreaCode.Dk1));
 
-            var userId = Guid.NewGuid();
+            var externalUserId = new ExternalUserId(Guid.NewGuid());
 
             var target = new GridAreaAuditLogEntryRepository(context);
-            var entry = new GridAreaAuditLogEntry(DateTimeOffset.UtcNow, userId, GridAreaAuditLogEntryField.Name, "old_val", "new_val", gridAreaId.Value);
+            var entry = new GridAreaAuditLogEntry(DateTimeOffset.UtcNow, externalUserId, GridAreaAuditLogEntryField.Name, "old_val", "new_val", gridAreaId);
 
             await target.InsertAsync(entry);
 
@@ -90,7 +91,7 @@ namespace Energinet.DataHub.MarketParticipant.IntegrationTests.Repositories
 
             // assert
             Assert.Single(actual);
-            Assert.Equal(gridAreaId.Value, actual.Single().GridAreaId);
+            Assert.Equal(gridAreaId, actual.Single().GridAreaId);
         }
     }
 }