+++ source/Energinet.DataHub.MarketParticipant.Client/documents/release-notes/release-notes.md
@@ -1,8 +1,16 @@
 # Energinet.DataHub.MarketParticipant.Client Release notes
 
-## Version 2.10.1
+## Version 2.12.1
 
 - Added optional searchText param to user overview
+
+## Version 2.12.0
+
+- Simplify call to assignment of user roles.
+
+## Version 2.11.0
+
+- Organization name for each actor in actor selection API.
 
 ## Version 2.10.0